Updated GitHub Action with time limit for agent job
In a prior deploy, the agent was stuck on the test step for nearly 6h. That is simply not acceptable.

From prior runs in the rails template application repository, we've had runs at less than 5 minutes. I'm doubling the limit to ensure that if something were to go awry, that a 5 minute limit isn't too restrictive.