#3 Refactored the OrdersController with utilities.
There was a lot of duplication regarding how the controller was handling errors that could occur as it was processing an order for shoes.

I made functions such as `handle_error_in_order` or `handle_valid_order` that will do basically was done before but it's more flexible in its application through the use of constants.